gp-import: supported double format of tempo for gp8

+++ src/importexport/guitarpro/internal/gtp/gpmastertracks.h
@@ -18,7 +18,7 @@ public:
         Type type{ Type::tempo };
         float position{ 0 };
         bool linear{ false };
-        int value{ 0 };
+        double value{ 0 };
         int tempoUnit{ 0 };
         muse::String text;
         friend bool operator<(const Automation& l, const Automation& r) { return l.bar < r.bar; }
